website/backend/requirements.txt
flask
requests
python-dotenv
flask-cors
pandas
boto3